Split the test into multiple files
